Stop reading full file into RAM when creating the zchunk file
File is full read into RAM before creating the zchunk file, and this
requires large amount of RAM. However, the procedure to create the ZCK
does not require to have the whole file in RAM, but just the chunk that
is worked and the header. In case autochunk mode is selected, read the
original file on demand and loads the chunk when they are required.
This allows to create zck for much larger files as in the past.

To avoid to break current API, a new function is introduced:

	zck_write_from_fd()

This accepts a file descriptor instead of the buffer where the whole
file was loaded.

Signed-off-by: Stefano Babic <sbabic@denx.de>