Fix size limit in zck.c
Up to now zck.c reads the whole file into memory and then sends it to
zck_write to actually compress.  This is highly inefficient, but was done
to simplify dealing with a string separator.  This commit fixes this by
reading the file in chunks, and, if using a string separator, keeps track
of the matching bytes from chunk to chunk.

Signed-off-by: Jonathan Dieter <jdieter@gmail.com>