Add ZCK_NO_WRITE option to disable writing to a file
There are situations where we need to recreate the zchunk header for an
uncompressed file without actually writing out the zchunk file.  This
commit allows setting an `ioption`, `ZCK_NO_WRITE`, that will go through
the full process of creating the zchunk file without actually writing it
to disk, leaving you with a complete zchunk header in the end.

Signed-off-by: Jonathan Dieter <jdieter@gmail.com>