Handle filenames passed to `unzck` more sanely
We were doing some very strange things when trying to figure out the output
filename, including failing if the input didn't end in `.zck` or `.zhr`.

None of this made sense, so now we're just stripping `.zck` from the input
file if it's there when calculating the output file name.

Thanks to Agostino Sarubbo of Gentoo for finding and reporting this.

Signed-off-by: Jonathan Dieter <jdieter@gmail.com>