Stop storing last error when missing zchunk context
Originally, when a null zchunk context was passed into the error logging
functions, we stored the error information in a global context called
`zck_none`.  While this worked perfectly for single-threaded operations,
multiple threads, even if operating on different files, would
intermittently fail (for an example, see
https://github.com/rpm-software-management/createrepo_c/issues/386).

This PR removes this "feature".  Now, when a null zchunk context is passed
to "create" the error, we output the error message, but don't do anything
else.  Running `zck_get_error(NULL)` will always return an empty string and
running `zck_clear_error(NULL)` is a noop.  `zck_is_error(NULL)` will
always return 0.

We also remove everywhere in our code that we were running
`zck_get_error(NULL)` and `zck_clear_error(NULL)`, since they're noops.

The reality is that the only way to not have a zchunk context is if we
weren't able to allocate it, so I expect this to have no effect on any
users of the library.

This makes the library thread-safe, as long as each context is only touched
in a single thread.

Signed-off-by: Jonathan Dieter <jonathan.dieter@cyara.com>